+++ bangazon-client/src/components/product/ProductCategory.js
@@ -13,6 +13,12 @@ const ProductCategory = props => {
     const getProducts = () => {
             fetch(`http://localhost:8000/products?category=${props.category.id}`, {
                 "method": "GET",
+                headers :{
+                    "Content-Type": "application/json",
+                    "Accept": "application/json",
+                    "Authorization": `Token ${localStorage.getItem("bangazon_token")}`,
+
+                }
             })
                 .then(response => response.json())
                 .then(setProducts)
