+++ bangazon-client/src/components/product/ProductForm.js
@@ -65,8 +65,8 @@ const ProductForm = props => {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
-        Accept: "application/json",
-        Authorization: `Token ${localStorage.getItem("bangazon_token")}`
+        "Accept": "application/json",
+        "Authorization": `Token ${localStorage.getItem("bangazon_token")}`
       },
       body: JSON.stringify(newProduct)
     }).then(res => res.json());
