+++ bangazon-client/src/components/product/ProductList.js
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { useEffect, useState } from "react"
 import Product from "./Product"
 
 // // """
@@ -8,13 +8,32 @@ import Product from "./Product"
 
 
 const ProductList = props => {
+    const [products, setProducts] = useState([])
+
+    const getProducts = () => {
+            fetch(`http://localhost:8000/products`, {
+                "method": "GET",
+                headers :{
+                    "Content-Type": "application/json",
+                    "Accept": "application/json",
+                    "Authorization": `Token ${localStorage.getItem("bangazon_token")}`,
+                }
+            })
+                .then(response => response.json())
+                .then(setProducts)
+        }
+
+
+    useEffect(() => {
+        getProducts()
+    }, [])
 
     return (
         <>
                 <ol>
             <article className="productList">
                 {
-                    props.products.slice(-20).map(product =>
+                    products.slice(-20).map(product =>
                         <Product key={product.id}
                             product={product} {...props} />)
                 }
